Update README.md
I edited the game description and "getting started" instructions for enhanced clarity and conciseness. Furthermore, I added the line "Download Python if you do not already have Python installed" for people who may not know what Python is and to indicate that python is not a default program that is natively installed. 

Please consider and accept my edits if you feel that the edits are good for your game. 

Thank You!